splitted into two files the tests streamflow.cwl.* classes from the others
